1：修改获取验证码接口

+++ src/main/java/com/sftc/web/dao/jpa/UserInviteDao.java
@@ -5,6 +5,6 @@ import com.sftc.web.model.UserInvite;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface UserInviteDao extends PagingAndSortingRepository<UserInvite, Long>, JpaSpecificationExecutor<UserInvite> {
+public interface UserInviteDao extends PagingAndSortingRepository<UserInvite, Integer>, JpaSpecificationExecutor<UserInvite> {
 
 }
